refactor: extract platform/pip.rs from process.rs
process.rs had accreted most of the crate's pip knowledge alongside its
actual subject, generic child-process supervision, and was brushing the
800-line cap. The pip-specific layer moves to a new platform/pip.rs:
the isolation env tables and helpers, pip_command, run_pip, the bounded
blocking install and its timeout, plus their tests. process.rs keeps
the generic machinery (bounded execution and drain, capture, python
isolation, window suppression, job objects) and lends it to pip.rs
through pub(super) seams; the env_edits test helper is hoisted to
module level so both test mods share it. Pure move, no behavior change;
callers outside platform are untouched because pip_command and run_pip
were already consumed through platform:: re-exports.

Fixes #362
